website/fitfriendsclub-deployment/backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
requests==2.32.4
PyJWT==2.9.0
python-jose>=3.4.0
Pillow>=10.3.0
gunicorn==21.2.0